Reworked exporters to be more functional in layout
